% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extract_formula}
\alias{extract_formula}
\title{Extract formula}
\usage{
extract_formula(dataframe)
}
\arguments{
\item{dataframe}{Dataframe of type data.frame. Must contain outcome and all observables.}

\item{outcome}{Terminal outcome of type character.}
}
\description{
Extract formula
}
\keyword{Dynamic}
\keyword{regime.}
\keyword{treatment}
