% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_policy.R
\name{optimal_policy}
\alias{optimal_policy}
\title{Optimal Policy}
\usage{
optimal_policy(exposure, exposure_lag, model, dataframe, init_val_optim = 1000)
}
\arguments{
\item{exposure}{Exposure of type character.}

\item{model}{Model object.}

\item{dataframe}{Dataframe of type data.frame.}
}
\description{
Optimal Policy
}
\keyword{Dynamic}
\keyword{regime.}
\keyword{treatment}
