% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalDTR.R
\name{optimalDTR}
\alias{optimalDTR}
\title{Optimal Treatment Regime}
\usage{
optimalDTR(
  exposure,
  covariates,
  outcome,
  stage_column,
  start_stage,
  end_stage,
  dataframe
)
}
\arguments{
\item{exposure}{Exposure variable. Type character.}

\item{covariates}{List of covariates of type characters.}

\item{outcome}{Terminal outcome. Type character.}

\item{stage_column}{Column to indicate stage of treatment. Type character.}

\item{start_stage}{Value to indicate first stage of treatment. Type integer}

\item{end_stage}{Value to indicate last stage of treatment. Type integer}

\item{dataframe}{Dataframe of type data.frame.}
}
\description{
Optimal Treatment Regime
}
\keyword{Dynamic}
\keyword{regime.}
\keyword{treatment}
