% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{stage_regression}
\alias{stage_regression}
\title{Stage Regression Function}
\usage{
stage_regression(dataframe)
}
\arguments{
\item{dataframe}{Dataframe of type data.frame. Must contain outcome and all observables.}

\item{outcome}{Terminal outcome of type character.}
}
\description{
Stage Regression Function
}
\keyword{Dynamic}
\keyword{regime.}
\keyword{treatment}
